+++ workspace/solutions/rust/src/s0/s7/s1/s3/mod.rs
@@ -1,0 +1,4 @@
+pub mod s_0713_00;
+pub mod s_0713_test;
+
+pub type TSolution = fn(Vec<i32>, i32) -> i32;