+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 更改日志 | Change Log
 
+## 2024-09-06
+
+- 添加题目 `0713` 的题解 | Add the solution to the question `0713`
+  - [713. 乘积小于 K 的子数组 - 力扣（LeetCode）](https://leetcode.cn/problems/subarray-product-less-than-k/description/)
+
 ## 2024-09-05
 
 - 添加题目 `2132` 的题解 | Add the solution to the question `2132`